Rename routes dir -> views
