+++ app/factory.py
@@ -5,7 +5,7 @@ from typing import List
 import json
 
 from app.extensions import db, migrate, rq, limiter, csrf, assets, cache, talisman
-from app.routes import base, auth, raffles, api, users
+from app.views import base, auth, raffles, api, users
 from app.config import ProdConfig
 from app.commands import delete, clear_cache
 from app.logging import configure_logger
