+++ app/views/raffles.py
